[DEV] Use the shorter integral type names across code

+++ include/qg_generator.hpp
@@ -1,16 +1,16 @@
 #pragma once
 
-#include <cstdint>
-
 #include <unordered_map>
 #include <string>
 #include <vector>
 
+#include "qg_types.hpp"
+
 // CASE GENERATOR ====================
 
-enum city_size : int8_t { SIZE_SMALL, SIZE_MEDIUM, SIZE_LARGE, SIZE_METRO, SIZE_COUNT };
-enum district_type : int8_t { INDUSTRIAL, COMMERCIAL, RESIDENTIAL, NIGHT_LIFE, FINANCIAL, DISTRICT_COUNT };
-enum landmark_type : int8_t {
+enum city_size : i8 { SIZE_SMALL, SIZE_MEDIUM, SIZE_LARGE, SIZE_METRO, SIZE_COUNT };
+enum district_type : i8 { INDUSTRIAL, COMMERCIAL, RESIDENTIAL, NIGHT_LIFE, FINANCIAL, DISTRICT_COUNT };
+enum landmark_type : i8 {
     //TODO: More variety
     // categorize by district type
     LOCATION_RESIDENTIAL,
@@ -20,30 +20,30 @@ enum landmark_type : int8_t {
     LOCATION_COUNT,
 };
 
-extern int8_t size_to_districts[city_size::SIZE_COUNT];
-extern int32_t district_weights[district_type::DISTRICT_COUNT];
-extern int8_t size_to_landmarks[city_size::SIZE_COUNT];
+extern i8 size_to_districts[city_size::SIZE_COUNT];
+extern i32 district_weights[district_type::DISTRICT_COUNT];
+extern i8 size_to_landmarks[city_size::SIZE_COUNT];
 
 struct district {
-    int64_t id;
+    i64 id;
     district_type type;
 };
 
 struct landmark {
-    int64_t id;
-    int64_t district;
+    i64 id;
+    i64 district;
 };
 
 struct case_gen {
     city_size size;
-    int8_t num_districts;
+    i8 num_districts;
     district districts[16];
 
-    int8_t num_landmarks;
+    i8 num_landmarks;
     landmark landmarks[128];
 };
 
-void case_gen_fondation(case_gen *ctx, city_size s, int32_t seed);
+void case_gen_fondation(case_gen *ctx, city_size s, i32 seed);
 void case_gen_population(case_gen *ctx);
 void case_gen_motive(case_gen *ctx);
 void case_gen_crime(case_gen *ctx);
@@ -57,14 +57,14 @@ void case_gen_polish(case_gen *ctx);
 //TODO: Find a way to simplify this data structure; remove maps/vectors
 //TODO: Each name_gen should have it's own memory pool for names
 struct name_gen {
-    std::unordered_map<std::string, std::unordered_map<char, int32_t>> counts;
+    std::unordered_map<std::string, std::unordered_map<char, i32>> counts;
 };
 extern const char *district_prefix[];
-extern const size_t num_district_prefix;
+extern const u64 num_district_prefix;
 extern const char *district_suffix[];
-extern const size_t num_district_suffix;
+extern const u64 num_district_suffix;
 
 void name_gen_train(name_gen *gen, const char *file_path);
-void name_gen_next(name_gen *gen, size_t num, std::vector<std::string> *out);
+void name_gen_next(name_gen *gen, u64 num, std::vector<std::string> *out);
 
-void name_gen_district(name_gen *gen, size_t num, std::vector<std::string> *out);
+void name_gen_district(name_gen *gen, u64 num, std::vector<std::string> *out);
